improved unit tests structure and removed all todos
