+++ backend/tests/entity_factories.py
@@ -1,4 +1,3 @@
-from modules.extract.domain import value_objects
 from modules.transform.domain import value_objects as transform_value_objects
 
 from . import fakers
@@ -11,16 +10,6 @@ class BatchMixin:
             raise ValueError("Size must be a positive integer")
 
         return [cls.create(**kwargs) for _ in range(size)]  # type: ignore[attr-defined]
-
-
-class ExtractHistoryFactory(BatchMixin):
-    @staticmethod
-    def create(**kwargs) -> value_objects.ExtractHistory:
-        return value_objects.ExtractHistory(
-            input_file_name=kwargs.get("input_file_name", fakers.fake_file_name()),
-            saved_file_name=kwargs.get("saved_file_name", fakers.fake_file_name()),
-            timestamp=kwargs.get("timestamp", fakers.fake_timestamp()),
-        )
 
 
 class TransformedDataFactory(BatchMixin):
