+++ backend/tests/fakers.py
@@ -2,15 +2,12 @@ import random
 import string
 from datetime import datetime
 
-from modules.transform.domain import value_objects as transform_value_objects
-
 from modules.extract.domain import value_objects as extract_value_objects
+from modules.transform.domain import value_objects as transform_value_objects
 
 
 def fake_name(length=10):
-    return "".join(
-        random.choice(string.ascii_letters) for _ in range(length)
-    )
+    return "".join(random.choice(string.ascii_letters) for _ in range(length))
 
 
 def fake_age():
@@ -36,7 +33,10 @@ def fake_transformed_data() -> transform_value_objects.TransformedData:
         full_name=fake_name(10), age=fake_age(), is_satisfied=fake_is_satisfied()
     )
 
+
 def fake_extract_history() -> extract_value_objects.ExtractHistory:
     return extract_value_objects.ExtractHistory(
-        input_file_name=fake_file_name(10), saved_file_name=fake_file_name(), timestamp=fake_timestamp()
+        input_file_name=fake_file_name(10),
+        saved_file_name=fake_file_name(),
+        timestamp=fake_timestamp(),
     )
