--- backend/tests/integrated/test_integration.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-    pass
